+++ src/test/java/edu/kis/vh/nursery/RhymersDemo.java
@@ -1,6 +1,5 @@
 package edu.kis.vh.nursery;
 
-import edu.kis.vh.nursery.factory.DefaultRhymersFactory;
 import edu.kis.vh.nursery.factory.Rhymersfactory;
 
 class RhymersDemo {
@@ -32,6 +31,29 @@ class RhymersDemo {
 
 		System.out.println("total rejected is "
 				+ ((HanoiRhymer) rhymers[3]).reportRejected());
+	}
+
+}
+class DefaultRhymersFactory implements Rhymersfactory {
+
+	@Override
+	public DefaultCountingOutRhymer getStandardRhymer() {
+		return new DefaultCountingOutRhymer();
+	}
+
+	@Override
+	public DefaultCountingOutRhymer getFalseRhymer() {
+		return new DefaultCountingOutRhymer();
+	}
+
+	@Override
+	public DefaultCountingOutRhymer getFIFORhymer() {
+		return new FIFORhymer();
+	}
+
+	@Override
+	public DefaultCountingOutRhymer getHanoiRhymer() {
+		return new HanoiRhymer();
 	}
 
 }
